test: move reestablish regression into channel test module
